Fix issue num steps DDP
The computed number of steps per gpu is wrong in DDP mode. As the sampler already divides the instances over the gpus and the batch size is also divided per gpu in the config file, it is not necessary to multiply it back to the number of gpus. The number of training step is computer for each gpu.

Related issue :
pairlab/SlotFormer#11